+++ CHANGELOG.md
@@ -21,6 +21,7 @@ All notable changes to this project will be documented in this file.
 - hbase: Add hadoop-azure.jar to the lib directory to support the Azure Blob Filesystem and
   the Azure Data Lake Storage ([#853]).
 - kafka: Add cyrus-sasl-gssapi package for kerberos ([#874]).
+- spark: Add HBase connector ([#878]).
 
 ### Changed
 
@@ -69,6 +70,7 @@ All notable changes to this project will be documented in this file.
 [#868]: https://github.com/stackabletech/docker-images/pull/868
 [#874]: https://github.com/stackabletech/docker-images/pull/874
 [#877]: https://github.com/stackabletech/docker-images/pull/877
+[#878]: https://github.com/stackabletech/docker-images/pull/878
 [#879]: https://github.com/stackabletech/docker-images/pull/879
 
 ## [24.7.0] - 2024-07-24